feat: Make id.test.ts work with array uniq ids
this works by cartesian producting our IDs, so that they're all added
to the set of unique IDs. If a chart has a unique ID On

["A", "B"], C, D

then A C D and B C D will be correctly be added to the set of uniq ids
now.

This is a regression test for the behaviour that let slightly broken
khamen break charts get into prod.